requirements.txt
requests==2.25.1
opensearch-py==1.1.0
spacy==3.3.0
tqdm==4.59.0
weaviate-client